- Scanner spectral path now stays double until the final write: LUT storage is std::vector<double> (Scanner.h), LUT build/sampling and per‑pixel spectral→logXYZ→XYZ use double math with a new double dyes→XYZ helper; CAT02 and XYZ→RGB multiply in double, and only the encoded RGB written to the host is cast to float (ScannerOptics.cpp/ScannerOptics.h). - Blur/unsharp now operate on double buffers with templated Gaussian kernel + separable blur to preserve the float64 pipeline while keeping glare math in float; kernels in the runtime are now std::vector<double> (ScannerOptics.cpp/ScannerOptics.h). - Output encoding gained a double overload so CCTF and optional DWG→output transforms run in double when invoked from the scanner; existing float behavior is unchanged (OutputEncoding.cpp/OutputEncoding.h). - Log10 epsilon remains 1e‑10 but applied in double; baseline handling mirrors agx via the new double dyes→XYZ helper (no hash/key changes).

+++ Scanner.h
@@ -113,7 +113,7 @@ namespace Scanner {
     };
 
     struct SpectralLutBuffer {
-        std::vector<float> cpu;
+        std::vector<double> cpu;
         std::uint64_t hash = 0;
         std::uint32_t res = 0;
         bool valid = false;
